Extract getting "lang pattern" into separate function to avoid code duplication
