+++ src/Presentation/Nop.Web/Infrastructure/BaseRouteProvider.cs
@@ -1,0 +1,27 @@
+﻿using System.Linq;
+using Microsoft.AspNetCore.Routing;
+using Microsoft.Extensions.DependencyInjection;
+using Nop.Core.Domain.Localization;
+using Nop.Data;
+using Nop.Services.Localization;
+
+namespace Nop.Web.Infrastructure
+{
+    public class BaseRouteProvider
+    {
+        protected string GetRouterPattern(IEndpointRouteBuilder endpointRouteBuilder, string seoCode = "")
+        {
+            if (DataSettingsManager.DatabaseIsInstalled)
+            {
+                var localizationSettings = endpointRouteBuilder.ServiceProvider.GetRequiredService<LocalizationSettings>();
+                if (localizationSettings.SeoFriendlyUrlsForLanguagesEnabled)
+                {
+                    var langservice = endpointRouteBuilder.ServiceProvider.GetRequiredService<ILanguageService>();
+                    var languages = langservice.GetAllLanguages().ToList();
+                    return "{language:lang=" + languages.FirstOrDefault().UniqueSeoCode + $"}}/{seoCode}";
+                }
+            }
+            return seoCode;
+        }
+    }
+}
