+++ src/Presentation/Nop.Web/Infrastructure/GenericUrlRouteProvider.cs
@@ -1,10 +1,5 @@
-﻿using System.Linq;
-using Microsoft.AspNetCore.Builder;
+﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Routing;
-using Microsoft.Extensions.DependencyInjection;
-using Nop.Core.Domain.Localization;
-using Nop.Data;
-using Nop.Services.Localization;
 using Nop.Web.Framework.Mvc.Routing;
 
 namespace Nop.Web.Infrastructure
@@ -12,7 +7,7 @@ namespace Nop.Web.Infrastructure
     /// <summary>
     /// Represents provider that provided generic routes
     /// </summary>
-    public partial class GenericUrlRouteProvider : IRouteProvider
+    public partial class GenericUrlRouteProvider : BaseRouteProvider, IRouteProvider
     {
         #region Methods
 
@@ -22,17 +17,8 @@ namespace Nop.Web.Infrastructure
         /// <param name="endpointRouteBuilder">Route builder</param>
         public void RegisterRoutes(IEndpointRouteBuilder endpointRouteBuilder)
         {
-            var pattern = "{SeName}";
-            if (DataSettingsManager.DatabaseIsInstalled)
-            {
-                var localizationSettings = endpointRouteBuilder.ServiceProvider.GetRequiredService<LocalizationSettings>();
-                if (localizationSettings.SeoFriendlyUrlsForLanguagesEnabled)
-                {
-                    var langservice = endpointRouteBuilder.ServiceProvider.GetRequiredService<ILanguageService>();
-                    var languages = langservice.GetAllLanguages().ToList();
-                    pattern = "{language:lang=" + languages.FirstOrDefault().UniqueSeoCode + "}/{SeName}";
-                }
-            }
+            var pattern = GetRouterPattern(endpointRouteBuilder, seoCode: "SeName");
+
             endpointRouteBuilder.MapDynamicControllerRoute<SlugRouteTransformer>(pattern);
 
             //and default one
@@ -47,25 +33,25 @@ namespace Nop.Web.Infrastructure
                 new { controller = "Common", action = "GenericUrl" });
 
             //define this routes to use in UI views (in case if you want to customize some of them later)
-            endpointRouteBuilder.MapControllerRoute("Product", pattern, 
+            endpointRouteBuilder.MapControllerRoute("Product", pattern,
                 new { controller = "Product", action = "ProductDetails" });
 
-            endpointRouteBuilder.MapControllerRoute("Category", pattern, 
+            endpointRouteBuilder.MapControllerRoute("Category", pattern,
                 new { controller = "Catalog", action = "Category" });
 
-            endpointRouteBuilder.MapControllerRoute("Manufacturer", pattern, 
+            endpointRouteBuilder.MapControllerRoute("Manufacturer", pattern,
                 new { controller = "Catalog", action = "Manufacturer" });
 
-            endpointRouteBuilder.MapControllerRoute("Vendor", pattern, 
+            endpointRouteBuilder.MapControllerRoute("Vendor", pattern,
                 new { controller = "Catalog", action = "Vendor" });
-            
-            endpointRouteBuilder.MapControllerRoute("NewsItem", pattern, 
+
+            endpointRouteBuilder.MapControllerRoute("NewsItem", pattern,
                 new { controller = "News", action = "NewsItem" });
 
-            endpointRouteBuilder.MapControllerRoute("BlogPost", pattern, 
+            endpointRouteBuilder.MapControllerRoute("BlogPost", pattern,
                 new { controller = "Blog", action = "BlogPost" });
 
-            endpointRouteBuilder.MapControllerRoute("Topic", pattern, 
+            endpointRouteBuilder.MapControllerRoute("Topic", pattern,
                 new { controller = "Topic", action = "TopicDetails" });
 
             //product tags
